--- recommend/html/section.html
@@ -1,27 +1,0 @@
-<div class="div-main">
-    <br>
-    <div class="v38_6837"></div>
-    <div class="wrapper">
-        {{if .GenericAdmissionControllerPolicy}}
-            <h3 class="v38_7030" align="center">Generic Kyverno Policies</h3>
-        {{else}}
-            <table align=center border=0 cellpadding=2>
-                {{range .ImgInfo}}
-                <tr align=left>
-                    <td>{{.Key}}</td>
-                    <td>: {{.Val}}</td>
-                </tr>
-                {{end}}
-            </table>
-        {{end}}
-    </div>
-    <div class="div-main">
-        <br>
-    </div>
-    <table border=0 id="table_detail" align=center cellpadding=30 >
-        <tr>
-            {{range .HdrCols}}
-                <th>{{.Name}}</th>
-            {{end}}
-        </tr>
-
