+++ recommend/image/yaml/distro.yaml
