+++ recommend/report/report_text.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2022 Authors of KubeArmor
 
-package recommend
+package report
 
 import (
 	_ "embed" // need for embedding
@@ -9,7 +9,8 @@ import (
 	"os"
 	"strings"
 
-	"github.com/accuknox/auto-policy-discovery/src/types"
+	"github.com/kubearmor/kubearmor-client/recommend/common"
+	"github.com/kubearmor/kubearmor-client/recommend/image"
 	"github.com/olekukonko/tablewriter"
 	log "github.com/sirupsen/logrus"
 )
@@ -30,7 +31,7 @@ func NewTextReport() TextReport {
 	}
 }
 
-func (r TextReport) writeImageSummary(img *ImageInfo) {
+func (r TextReport) writeImageSummary(img *image.Info, outDir string, currentVersion string) {
 	t := tablewriter.NewWriter(r.outString)
 	t.SetBorder(false)
 	if img.Deployment != "" {
@@ -41,26 +42,14 @@ func (r TextReport) writeImageSummary(img *ImageInfo) {
 	t.Append([]string{"OS", img.OS})
 	t.Append([]string{"Arch", img.Arch})
 	t.Append([]string{"Distro", img.Distro})
-	t.Append([]string{"Output Directory", img.getPolicyDir()})
-	t.Append([]string{"policy-template version", CurrentVersion})
+	t.Append([]string{"Output Directory", img.GetPolicyDir(outDir)})
+	t.Append([]string{"policy-template version", currentVersion})
 	t.Render()
 }
 
 // Start Start of the section of the text report
-func (r TextReport) Start(img *ImageInfo) error {
-	r.writeImageSummary(img)
-	r.table.SetHeader([]string{"Policy", "Short Desc", "Severity", "Action", "Tags"})
-	r.table.SetAlignment(tablewriter.ALIGN_LEFT)
-	r.table.SetRowLine(true)
-	return nil
-}
-
-func (r TextReport) StartGenericAdmissionControllerPolicies() error {
-	t := tablewriter.NewWriter(r.outString)
-	t.SetBorder(false)
-	t.Rich([]string{"Generic Kyverno Policies"}, []tablewriter.Colors{{tablewriter.Bold}})
-	t.Render()
-
+func (r TextReport) Start(img *image.Info, outDir string, currentVersion string) error {
+	r.writeImageSummary(img, outDir, currentVersion)
 	r.table.SetHeader([]string{"Policy", "Short Desc", "Severity", "Action", "Tags"})
 	r.table.SetAlignment(tablewriter.ALIGN_LEFT)
 	r.table.SetRowLine(true)
@@ -76,7 +65,7 @@ func (r TextReport) SectionEnd() error {
 }
 
 // Record addition of new text table row
-func (r TextReport) Record(ms MatchSpec, policyName string) error {
+func (r TextReport) Record(ms common.MatchSpec, policyName string) error {
 	var rec []string
 	policyName = policyName[strings.LastIndex(policyName, "/")+1:]
 	rec = append(rec, wrapPolicyName(policyName, 35))
@@ -84,19 +73,6 @@ func (r TextReport) Record(ms MatchSpec, policyName string) error {
 	rec = append(rec, fmt.Sprintf("%d", ms.Spec.Severity))
 	rec = append(rec, string(ms.Spec.Action))
 	rec = append(rec, strings.Join(ms.Spec.Tags[:], "\n"))
-	r.table.Append(rec)
-	return nil
-}
-
-// RecordAdmissionController adds new row to table for admission controller policies
-func (r TextReport) RecordAdmissionController(policyFilePath, action string, annotations map[string]string) error {
-	var rec []string
-	policyFilePath = policyFilePath[strings.LastIndex(policyFilePath, "/")+1:]
-	rec = append(rec, wrapPolicyName(policyFilePath, 35))
-	rec = append(rec, annotations[types.RecommendedPolicyTitleAnnotation])
-	rec = append(rec, "-")
-	rec = append(rec, action)
-	rec = append(rec, strings.Join(strings.Split(annotations[types.RecommendedPolicyTagsAnnotation], ",")[:], "\n"))
 	r.table.Append(rec)
 	return nil
 }
